Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/dom/nodes/ProcessingInstruction-literal-1-expected.txt


PASS <?xml?> is not a ProcessingInstruction
